fix: centralize game processing
