--- frontend/static/quarantine.json
@@ -1,8 +1,0 @@
-[
-    "401411157",
-    "401403861",
-    "401628329",
-    "401634301",
-    "401634212",
-    "401628398"
-]
